fix create request error handler

+++ root/Admin/include/update-user.php
@@ -10,42 +10,72 @@ if (isset($_POST['updateBtn'])) {
     $address = trim($_POST["address"]);
     $Image = $_FILES['profileImg']['name'];
     if ($Image == null) {
-        $sql = "UPDATE adduser set firstname=?,lastname=?,position=?,email=?,address=? where uID=?";
-        $stmt = $conn->prepare($sql);
         try {
-            if (!$stmt) {
-                throw new Exception('There was a problem executing the query.');
+            $updateUser = $conn->prepare("UPDATE adduser set firstname=?,lastname=?,position=?,email=?,address=? where uID=?");
+            $updateUser->bind_param("sssssi", $fname, $lname, $position, $email, $address, $uID);
+            if (!$updateUser->execute()) {
+                throw new Exception('There was a problem executing the query' . $conn->error);
             } else {
-                $stmt->bind_param("sssssi", $fname, $lname, $position, $email, $address, $uID);
-                if (!$stmt->execute()) {
-                    throw new Exception('There was a problem executing the query.');
-                } else {
-                    echo "success";
-                }
+                $response = array(
+                    'status' => 'Success',
+                    'message' => 'Your profile has been updated successfully',
+                    'icon' => 'success'
+                );
+                header('Content-Type: application/json');
+                echo json_encode($response);
+                exit();
             }
         } catch (Exception $e) {
             if ($e->getCode() == 1062) {
-                // Handle duplicate email error
-                echo "Email already exists";
+                $response = array(
+                    'status' => 'Error',
+                    'message' => 'Email address already exist',
+                    'icon' => 'error',
+                    'duplication' => false
+                );
+                header('Content-Type: application/json');
+                echo json_encode($response);
+                exit();
             } else {
-                // Handle other errors
-                echo $e->getMessage();
+                $response = array(
+                    'status' => 'Error',
+                    'message' => $e->getMessage(),
+                    'icon' => 'error',
+                );
+                header('Content-Type: application/json');
+                echo json_encode($response);
+                exit();
             }
         }
     } else {
-        $selectOldImg = "SELECT profile from adduser where uID=?";
-        $stmt = $conn->prepare($selectOldImg);
-        $stmt->bind_param('i', $uID);
-        $stmt->execute();
-        $oldRes = $stmt->get_result();
-        $old = $oldRes->fetch_assoc();
-        $oldImg = $old['profile'];
-        $path = "../include/profile/" . $oldImg;
-        if ($oldImg != null) {
-            if (file_exists($path)) {
-                unlink($path);
+        $selectOldImg = $conn->prepare("SELECT profile from adduser where uID=?");
+        $selectOldImg->bind_param('i', $uID);
+        $selectOldImg->execute();
+        try {
+            $oldRes = $selectOldImg->get_result();
+            if ($oldRes->num_rows < 0) {
+                throw new Exception("Failed to fetch data from database" . $conn->error);
+            } else {
+                $old = $oldRes->fetch_assoc();
+                $oldImg = $old['profile'];
+                $path = "../include/profile/" . $oldImg;
+                if ($oldImg != null) {
+                    if (file_exists($path)) {
+                        unlink($path);
+                    }
+                }
             }
+        } catch (Exception $e) {
+            $response = array(
+                'status' => 'Error',
+                'message' => $e->getMessage(),
+                'icon' => 'error',
+            );
+            header('Content-Type: application/json');
+            echo json_encode($response);
+            exit();
         }
+
         /*****************Update account*********************************/
 
         /*****************Upload new image**********************************/
@@ -57,29 +87,57 @@ if (isset($_POST['updateBtn'])) {
         if (move_uploaded_file($_FILES['profileImg']['tmp_name'], $filePath . $filename)) {
             if ($fileError === 0) {
                 if ($fileSize < 1000000) {
-                    $sql = "UPDATE adduser set firstname=?,lastname=?,position=?,email=?,address=?,profile=? where uID=?";
-                    $stmt = $conn->prepare($sql);
                     try {
-                        if (!$stmt) {
-                            throw new Exception('There was a problem executing the query.');
+                        $updateUserWithImage = $conn->prepare("UPDATE adduser set firstname=?,lastname=?,position=?,email=?,address=?,profile=? where uID=?");
+                        $updateUserWithImage->bind_param("ssssssi", $fname, $lname, $position, $email, $address, $filename, $uID);
+                        if (!$updateUserWithImage->execute()) {
+                            throw new Exception('There was a problem executing the query' . $conn->error);
                         } else {
-                            $stmt->bind_param("ssssssi", $fname, $lname, $position, $email, $address, $filename, $uID);
-                            if (!$stmt->execute()) {
-                                throw new Exception('There was a problem executing the query.');
+                            $getNewProfile = $conn->prepare("SELECT profile from adduser where uID=?");
+                            $getNewProfile->bind_param("i", $uID);
+                            if (!$getNewProfile->execute()) {
+                                throw new Exception('There was a problem executing the query' . $conn->error);
                             } else {
-                                echo "success";
+                                $getNewProfileResult = $getNewProfile->get_result();
+                                if ($getNewProfileResult->num_rows < 0) {
+                                    throw new Exception('Failed to fetch data from database');
+                                } else {
+                                    $newProfile = $getNewProfileResult->fetch_assoc();
+                                    $newFetchedProfile = $newProfile['profile'];
+                                    $response = array(
+                                        'status' => 'Success',
+                                        'message' => 'Your profile has been updated successfully',
+                                        'data' => $newFetchedProfile,
+                                        'icon' => 'success'
+                                    );
+                                    header('Content-Type: application/json');
+                                    echo json_encode($response);
+                                    exit();
+                                }
                             }
                         }
                     } catch (Exception $e) {
                         if ($e->getCode() == 1062) {
-                            // Handle duplicate email error
-                            echo "Error: Email already exists";
+                            $response = array(
+                                'status' => 'Error',
+                                'message' => 'Email address already exist',
+                                'icon' => 'error',
+                                'duplication' => false
+                            );
+                            header('Content-Type: application/json');
+                            echo json_encode($response);
+                            exit();
                         } else {
-                            // Handle other errors
-                            echo $e->getMessage();
+                            $response = array(
+                                'status' => 'Error',
+                                'message' => $e->getMessage(),
+                                'icon' => 'error',
+                            );
+                            header('Content-Type: application/json');
+                            echo json_encode($response);
+                            exit();
                         }
                     }
-                    
                 }
             }
         }
@@ -92,26 +150,42 @@ if (isset($_POST['updatePassword'])) {
     $passUID = $_POST['uID'];
     $currentPass = $_POST['currentPass'];
     $newPass = $_POST['newPass'];
-
-    $pass = "SELECT pwdUsers from adduser where uID= ?";
-    $stmt = $conn->prepare($pass);
-    $stmt->bind_param('i', $passUID);
-    $stmt->execute();
-    $result = $stmt->get_result();
-    while ($pw = $result->fetch_assoc()) {
-        $hash = $pw['pwdUsers'];
-        if (password_verify($currentPass, $hash)) {
-            $updatePass = "UPDATE adduser set pwdUsers=? where uID=?";
-            $stmt = $conn->prepare($updatePass);
-            $hashPW = password_hash($newPass, PASSWORD_DEFAULT);
-            $stmt->bind_param('si', $hashPW, $passUID);
-            $stmt->execute();
-            echo "success";
-            $stmt->close();
+    try {
+        $getCurrentPassword =  $conn->prepare("SELECT pwdUsers from adduser where uID= ?");
+        $getCurrentPassword->bind_param('i', $passUID);
+        if (!$getCurrentPassword->execute()) {
+            throw new Exception('There was a problem executing the query' . $conn->error);
         } else {
-            echo "Password does not match to any account";
+            $getCurrentPasswordResult = $getCurrentPassword->get_result();
+            while ($pw = $getCurrentPasswordResult->fetch_assoc()) {
+                $hash = $pw['pwdUsers'];
+                if (password_verify($currentPass, $hash)) {
+                    $updateUserPassword = $conn->prepare("UPDATE adduser set pwdUsers=? where uID=?");
+                    $hashPW = password_hash($newPass, PASSWORD_DEFAULT);
+                    $updateUserPassword->bind_param('si', $hashPW, $passUID);
+                    $updateUserPassword->execute();
+                    $response = array(
+                        'status' => 'Success',
+                        'message' => 'Your password is successfully changed',
+                        'icon' => 'success',
+                    );
+                    header('Content-Type: application/json');
+                    echo json_encode($response);
+                    exit();
+                } else {
+                    throw new Exception("Password does not match to any account");
+                }
+            }
         }
+    } catch (Exception $e) {
+        $response = array(
+            'status' => 'Error',
+            'message' => $e->getMessage(),
+            'icon' => 'error',
+        );
+        header('Content-Type: application/json');
+        echo json_encode($response);
+        exit();
     }
-    $conn->close();
 }
 /*****************Update Password**********************************/
